Make cppcheck on ubuntu1404.32bit happy
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@18157 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ tracking/trackFindingCDC/trackFinderOutputCombining/include/FittingMatrix.h
@@ -53,7 +53,7 @@ namespace Belle2 {
        */
       void resetSegment(SegmentCounter counterSegment)
       {
-        for (TrackCounter counterTracks = 0; counterTracks < m_fittingMatrix.cols(); counterTracks++) {
+        for (TrackCounter counterTracks = 0; counterTracks < static_cast<unsigned int>(m_fittingMatrix.cols()); counterTracks++) {
           resetEntry(counterSegment, counterTracks);
         }
       }
